==> 00_Análisis_API.ipynb <==
# <span style="font-size: 1.5em;">Análisis de la API de AGN</span> 

Con el objetivo de analizar los informes de auditorías disponibles en la página web de la Auditoría General de la Nación (AGN), es esencial comenzar por entender la estructura y funcionalidad de la API asociada a esta página. La AGN cuenta con una sección específica dedicada a las búsquedas de auditorías, la cual puede ser accedida a través del enlace: https://www.agn.gob.ar/auditorias/buscador. Se identifico este segmento del sitio web realiza peticiones a la siguiente dirección de API: https://webagnapi.agn.gob.ar/api.

En las subsecuentes secciones, se desglosaran las diversas solicitudes y respuestas relacionadas con esta API. Analizando a profundidad las principales características y los datos que esta herramienta nos permite acceder.

#### <span style="font-size: 1.5em;">Librerías</span>

import requests
# Permite hacer solicitudes HTTP a servidores web o APIs de manera sencilla.
import json
# Facilita la codificación y decodificación de datos en formato JSON.

#### <span style="font-size: 1.5em;">Contenido de la API</span> 

# Definimos la URL de la API que queremos analizar
url = "https://webagnapi.agn.gob.ar/api"

# Realizamos una solicitud GET a la API. Desactivamos la verificación SSL con verify=False (esto puede ser un riesgo de seguridad)
response = requests.get(url, verify=False)

# Verificamos si la respuesta tiene un código de estado 200, lo que indica una solicitud exitosa
if response.status_code == 200:
    # Convertimos la respuesta en formato JSON a un objeto Python
    data = response.json()
    
    # Extraemos la información de los enlaces (links) de la respuesta
    links = data.get('links', {})
    
    # Creamos una lista de URLs (hrefs) a partir de la información de los enlaces
    hrefs = [link_info.get('href') for link_info in links.values()]
    
    # Imprimimos cada URL con un número de orden
    for i, href in enumerate(hrefs, 1):
        print(f"{i}. {href}")
else:
    # Si la respuesta no tiene un código de estado 200, imprimimos un mensaje de error con el código de estado recibido
    print(f"La solicitud falló con el código de estado {response.status_code}")


#### <span style="font-size: 1.5em;">Exploracion de URLs</span> 

# Definimos la URL de la API que queremos analizar. Ejemplo:
url = "https://webagnapi.agn.gob.ar/api/node/infografias"

# Realizamos una solicitud GET a la URL especificada.
response = requests.get(url)

# Verificamos si la respuesta tiene un código de estado 200, lo que indica una solicitud exitosa.
if response.status_code == 200:
    # Convertimos la respuesta en formato JSON a un objeto Python.
    data = response.json()
    
    # Imprimimos el contenido de la respuesta en formato JSON con una indentación de 4 espacios y ordenando las claves.
    print(json.dumps(data, indent=4, sort_keys=True))
else:
    # Si la respuesta no tiene un código de estado 200, imprimimos un mensaje de error con el código de estado recibido.
    print("Error:", response.status_code)

Al explorar las distintas URLs, se identifico que se está utilizando el formato JSON:API, que es una especificación para la construcción de APIs utilizando JSON. Está diseñado para minimizar tanto el número de solicitudes como la cantidad de datos transferidos entre clientes y servidores, por lo que se tiene que tener ciertas consideraciones:

-	Limitación en Solicitudes: Las APIs pueden tener un límite en la cantidad de peticiones permitidas en un tiempo específico. Si excedes este límite, recibirás errores en tus peticiones.

-	Paginación: A menudo, las APIs retornan resultados en páginas para no sobrecargar la respuesta. Es posible que tengas que gestionar múltiples solicitudes para obtener todos los datos.

-	Profundidad de los Recursos: Si bien JSON:API permite recursos anidados, pueden limitar la profundidad de la respuesta para evitar respuestas excesivamente grandes.

-	Campos Específicos: Aunque es posible solicitar campos concretos, no todas las APIs soportan esta funcionalidad. Es posible que termines recibiendo más información de la deseada.

-	Errores de Validación: Solicitudes mal formadas o campos inexistentes pueden resultar en errores.

-	Actualización: Las APIs evolucionan, y lo que funcionó en un momento puede no funcionar después debido a cambios en la estructura o en los campos disponibles.


#### <span style="font-size: 1.5em;">Selección de URLs para la Creación de la Base de Datos</span> 

De las 107 URLs que componen la API de la AGN, se realizó una selección crítica para centrarse en aquellas que proporcionan la información más relevante para nuestro análisis y posterior creación de la base de datos. A continuación, se detallan las 10 URLs seleccionadas:

50.	https://webagnapi.agn.gob.ar/api/node/informes

51.	https://webagnapi.agn.gob.ar/api/node/infografias

52.	https://webagnapi.agn.gob.ar/api/file/file

53.	https://webagnapi.agn.gob.ar/api/taxonomy_term/jurisdiccion

54.	https://webagnapi.agn.gob.ar/api/taxonomy_term/objeto_de_auditoria

55.	https://webagnapi.agn.gob.ar/api/taxonomy_term/ods

56.	https://webagnapi.agn.gob.ar/api/taxonomy_term/organismo_auditado

57.	https://webagnapi.agn.gob.ar/api/taxonomy_term/sector_publico_nacional

58.	https://webagnapi.agn.gob.ar/api/taxonomy_term/tags

59.	https://webagnapi.agn.gob.ar/api/taxonomy_term/tipo_de_auditoria

#### <span style="font-size: 1.5em;">Contenido y estructura</span> 

Informes - La URL proporcionada ofrece detalles descriptivos de los informes, así como los datos empleados en los filtros del sitio web.

50.	https://webagnapi.agn.gob.ar/api/node/informes

Dentro de la estructura JSON de la URL, se distinguen dos componentes esenciales para cada informe: attributes y relationships. Estos componentes contienen información descriptiva y relacional de los informes, respectivamente.

Componente "attributes": Contiene los atributos descriptivos asociados a un informe, brindando detalles específicos sobre el mismo.

Componente "relationships": Representa las conexiones del informe con otras entidades o nodos en el sistema. Cada relación está definida por una clave única y ofrece información sobre la entidad relacionada.

A continuación, se detalla la estructura de estos componentes:

"data": [
    "attributes": {
        "titulo":
        "drupal_internal__nid":
        "cuerpo":
            "value": "<p>
        "alias":
        "ano":
        "archivo_difusion":
        "auditoria_coordinada":
        "material_de_difusion":
        "actuacion": {
            "value":
        "objetivo":
        "periodo_auditado": {
            "value":
            "end_value":
        "resolucion":
        "titulo_difusion":   
            
    "relationships": {
        "node_type": {
            "data": {
                "meta": {
                     "drupal_internal__target_id":
        "anexo": {
            "meta": {
                    "drupal_internal__target_id":
        "ficha": {
            "meta": {
                    "drupal_internal__target_id":
        "imagen": { 
            "meta": {
                    "drupal_internal__target_id":
        "infografia": {
            "meta": {
                    "drupal_internal__target_id":
        "informe": {
            "data": {
                "meta": {
                    "drupal_internal__target_id":
        "informe_en_video": {
            "meta": {
                    "drupal_internal__target_id":
        "informe_multimedia": { 
            "meta": {
                    "drupal_internal__target_id":
        "jurisdiccion": {
            "meta": {
                    "drupal_internal__target_id":
        "objeto_de_auditoria": {
            "meta": {
                    "drupal_internal__target_id":
        "ods": {
            "meta": {
                    "drupal_internal__target_id":
        "organismo_auditado": {
            "data": [
                "meta": {
                    "drupal_internal__target_id":
        "palabras_claves": {
            "meta": {
                    "drupal_internal__target_id":
        "resolucion_archivo": {
            "data": {
                "meta": {
                    "drupal_internal__target_id":
        "sector_publico_nacional": {
            "meta": {
                    "drupal_internal__target_id":
        "tipo_de_auditoria": {
            "meta": {
                    "drupal_internal__target_id":

Archivos - Esta URL sirve como fuente para los enlaces de archivos y documentos presentes en el sitio web.

52.	https://webagnapi.agn.gob.ar/api/file/file

Esta dirección está vinculada con la URL de informes a través del componente relationships bajo la clave "drupal_internal__fid".

Estructura JSON, destaca el componente "attributes":

"data": [
    "attributes": {
        "drupal_internal__fid":
        "filename":
        "uri": {
            "url":

Descripción de Filtros - Las siguientes URLs contienen categorías descriptivas que se utilizan como filtros en el sitio web de AGN:

53.	https://webagnapi.agn.gob.ar/api/taxonomy_term/jurisdiccion

55.	https://webagnapi.agn.gob.ar/api/taxonomy_term/ods

56.	https://webagnapi.agn.gob.ar/api/taxonomy_term/organismo_auditado

57.	https://webagnapi.agn.gob.ar/api/taxonomy_term/sector_publico_nacional

58.	https://webagnapi.agn.gob.ar/api/taxonomy_term/tags

59.	https://webagnapi.agn.gob.ar/api/taxonomy_term/tipo_de_auditoria

54.	https://webagnapi.agn.gob.ar/api/taxonomy_term/objeto_de_auditoria


Todas estas URLs presentan una estructura JSON similar, con una variación en el componente "attributes", que puede contener "name" o "nombre" como clave descriptiva. Además, están interconectadas con la URL de informes mediante el componente relationships a través del identificador "drupal_internal__tid".

Estructura JSON:


"data": [
    "type":
    "attributes": {
        "drupal_internal__tid":
        "name" o "nombre":

Infografías - La URL proporcionada alberga información adicional y archivos asociados a los informes.

51.	https://webagnapi.agn.gob.ar/api/node/infografias

La estructura JSON de esta dirección se compone de dos componentes principales: "attributes" y "relationships". Está vinculada a la URL de informes a través del identificador "drupal_internal__nid" en el componente "attributes". Además, se relaciona con la URL de archivos mediante el identificador "drupal_internal__target_id" en el componente "relationships".

Estructura JSON destacada:

"data": [
    "attributes": {
        "drupal_internal__nid":
        "title":
    "relationships": {
        "field_imagen":
            "data": [
                "meta": {
                    "alt":
                    "drupal_internal__target_id":
